chore: move 'fonts' into utils folder
