+++ .storybook/preview.tsx
@@ -2,7 +2,7 @@ import '../src/app/globals.css';
 import React from 'react';
 import { withThemeByClassName } from '@storybook/addon-styling';
 
-import { poppins } from '../src/lib/fonts';
+import { poppins } from '../src/lib/utils/fonts';
 
 import type { Preview } from '@storybook/react';
