+++ src/app/layout.tsx
@@ -1,5 +1,5 @@
 import './globals.css';
-import { lpmq, poppins } from '~/lib/fonts';
+import { lpmq, poppins } from '~/lib/utils/fonts';
 import { NavigationBar } from '~/components/molecules';
 import { ThemeProvider } from '~/components/organisms';
 import { cn } from '~/lib/utils/cn';
